requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
python-dateutil>=2.8.2
lxml>=4.9.0
playwright>=1.40.0
